Fix fastcore module bugs when running fastcore with no penalty set
In the case where the penalty set is empty the lp10 would simply
return the empty set (as an empty iterator), which is incorrect.
It could be argued that when the penalty set is empty it does not
even make sense to run fastcore but for consistency it will now
return the correct result even in this case.

In addition some comments were expanded and a function call typo
was corrected.