metabolicmodel: Run Fastcore on the induced reaction subset by Fastcc
A consistent subset is first obtained using Fastcc and this subset is
then expanded using Fastcore. This ensures that even if the original
model is not consistent the input to Fastcore will be. The biomass
reaction is also added to the Fastcore input since it is an error
if this reaction cannot be part of the model.