fastcore: Change fastcc to yield inconsistent set, add helper functions
fastcc() will yield all inconsistent reactions as soon as they are found.
This function is used to implement fastcc_is_consistent() which will
return False as soon as an inconsistent reaction is found, without
having to find all of them. The function fastcc_consistent_subset()
will return the consistent subset as a set.