fastcore: Add flipable model proxy wrapping MetabolicModel
This moves the ability to flip a model from the actual model into
a proxy object that provides a view of the model, modifying the
returned values such that it appears flipped. This means that
the MetabolicModel has a simpler implementation that is easier
to test, and the FlipableModelView can be tested separately. The
FlipableModelView can be applied by any algorithms that need this
kind of functionality.