metabolicmodel: Use default v_max value on model and store changes from this
This changes the limits property to a proxy object that provides a FluxBounds
instance for each reaction in the model. The FluxBounds is a proxy giving
access to the computed values of the bounds (based on the defaults but with
any explicitly set bound changed).

To clean up the code in the module, internal classes were moved to the top
level of the module. This allows the classes in FlipableModelView to inherit
from the base classes of MetabolicModel (LimitsView and
StoichiometricMatrixView). Also, these classes now inherit from the
Mapping ABC in the Python collections module.