reaction: Remove deprecated id property of Compound
It is no longer needed since the LP solver no longer needs string
ids for variable names (any hashable object can be used). In
addition the id property is not able to properly ensure uniqueness
of the id, and lastly it does not seem that Compound should be
responsible for generating unique ids in any case.