command: Add flux consistency check command
A reaction is flux consistent if there exists a steady-state flux
solution where the flux of the given reaction is non-zero. The
command will generate a list of reactions in the model that are
not flux consistent. By default, the fastcc check in the fastcore
module is used, but the naive consistency check implemented in
the fluxanalysis module is also available.