cplex: Redirect Cplex log output to Python logging
A LoggerFile class as added in the util module which implements a
file-like object for writing. Instances of this class can then be
handed to Cplex and the Cplex log output will be redirected through
the Python logging module.

This also changes the cplex.Solver() constructor since a stream
can no longer be given. To control the amount of debug output
from Cplex, set a logging level or a more specific logging filter.