expression: Change substitute to take a mapping function
This changes `substitute` in expressions to use a mapping
function instead of a dictionary (through kwargs). This works
better in practice since a default value cannot be given for
kwargs. Another option would be to take a dict object directly,
which could be an instance of defaultdict, thereby allowing default
values. This has the problem of not easily allowing a default of
mapping a variable to itself.