Increase robustness of misc reaction parsers
This changes the misc parsers in metnet.datasource.misc to more
robustly parse differences in reaction equations. Both functions
have added parameters to change the default reaction arrow
representation. The sudensimple parser is able to infer a compound
stoichiometry of one if the number is absent.

In addition the compartment of compounds is parsed using a regexp
pattern. This means that compound IDs can include multiple sets
of square brackets (only the last set will be interpreted as
compartment).

Lastly, the translation of compound IDs in metnet is abandoned
as the Compound class no longer imposes restrictions on the
compounds IDs.