Change internal model format to a central YAML file
This format allows specifying the model fully using one model file.
In addition, the model file can refer to other files to specify
parts of the model. The files in the subdirectory `reactions` will
be loaded by default (if existing) to define the reactions in the
model.

The command module is changed to only take the model file as input,
instead of having to take many files as input specifying different
aspects of the model. The command module will try to load the model
in the current directory if no model file was specified.