command: Change Command to take NativeModel as input
This changes the Command to have a constructor taking the
NativeModel and the argparse.Namespace. The base class Command
will construct a MetabolicModel from the NativeModel automatically.
This is needed for the individual commands to access the
NativeModel which can now be extended to include further model
information (e.g. list of Biomass reactions).