native: Parse compounds from either YAML or ModelSEED TSV
This changes the compounds key in the model specification to a list
of dict objects that represent individual compounds. In addition,
if the include key on the compound is set, the given file is
instead included. The format key can also be set, and must be set
if the format is not YAML.

Currently, only `modelseed` is recognized (in addition to `yaml`)
and will result in the compounds table being read as a ModelSEED
compound table.