Add generic Solver that automatically selects an LP solver
This adds a new module metnet.lpsolver.generic with a subclass of
metnet.lpsolver.lp.Solver that does not itself interface with a
solver but instead delegates to another solver based on specified
requirements. The generic interface knows about Cplex and QSopt_ex
and can decide on a solver based on whether integer values (MILP)
are required and whether a rational result is required.