sbml: Add non-strict parsing mode
Setting the parameter strict to False when creating the SBMLDatabase enables
a non-strict parsing mode which is based on the behavior of COBRA. Instead of
reporting ParseErrors it simply tries to ignore the errors or tries to parse
the file in a non-standard way. In addition it uses the postfix of the
compound ID to set the default value of boundaryCondition.