sbml: Change SBML parser to return ReactionEntries and CompoundEntries
This changes the SBML parser to simply return entries for reactions
and compounds instead of trying to implement the full Database
interface. This is needed to fit SBML import into model-import and it
makes the SBML parser code cleaner since SBML did not fit well into
the internal Database interface.