fastcore: Fix test of model with tiny stoichiometric coefficient
Corrected the value of the thresholds in the test description. Previously the
epsilon used in the test was exactly at the point between a failing and a
succeeding test. This meant that the uncertainty in the floating point
equality test could make the test fall to either side. The epsilon was lowered
to avoid this case.