datasource: Add FileMark tracking file and position of entries
Creates the context module in psamm.datasource containing FileMark and
also the FilePathContext class previously in native. The FileMark contains
a context, line and column, thereby keeping track of a position in a file.
The CompoundEntry and ReactionEntry of modelseed and native are extended to
include the filemark property which returns the position where the entry
was parsed.