command: Avoid randomsparse model copying by modifying constraints
Previously, when deleting reactions from the model a copy of the model was
made and the flux bounds in that copy were set to 0. This resulted in a new
LP problem being generated for each iteration. Using the new interface to
modifying constraints in the LP, a problem can be created once and the
constraints can be modified directly. This is much faster then creating a new
problem each time.