fluxanalysis: Merge into one FluxBalanceProblem object
Remove the FluxBalanceTDProblem and move the contents into the
FluxBalanceProblem as a method add_thermodynamic() that adds the
necessary variables and thermodynamic contraints. The solve() method is
renamed to maximize() and additional methods are added for minimize_l1()
and the convenience method max_min_l1(). The method minimize_l1() applies
the minimization of the L1 norm of fluxes previously found in
flux_minimization() and flux_minimization() is now just a wrapper. The
max_min_l1() method first uses maximize(), then contrains the maximized
reaction to the optimal value, and finally call minimize_l1().