fluxanalysis: Add functions to check for spurious production/consumption
Adds two functions that check for spurious production and for spurious
consumption, respectively. For example, the reactions A => B + C and C => A
allow for a spurious production of compound B. This check is complementary
to the stoichiometric check because the stoichiometric check reports on
potential model problems whereas the spurious loop check reports on
the subset of stoichiometric problems that actually influence the
simulation results.