robustness: Allow calculating flux ranges during robustness
This adds an option --fluxes to robustness which can be set to `all`,
`objective` or `range. `objective` works as the default mode before this
change. `all` works like the option `--all-reaction-fluxes` before (this
option is now gone). `range` calculates the minimum and maximum flux for
every model reaction with the objective reaction fixed at the maximum and
the varying reaction fixed at the flux of the iteration.