fluxanalysis: Improve stability with thermodynamic constraints
Reduce the allowed range of the delta mu variables in the LP problem with
thermodynamic constraints. This changes the epsilon to a hardcoded value
of one and the em parameter to a default of 1000.

These two variables constrained the delta mu variables of the reactions.
The epsilon was removed as a parameter and hardcoded to one because there
is no advantage in being able to adjust both parameters. In addition this
greatly reduced the range of the delta mu variables, solving a problem
where thermodynamic constraints would often lead to instability in the
LP problem which resulted in the solver reporting the problem as
infeasible. The documentation and comments have been updated to better
describe the variables and the issues with the em parameter.

This also fixes #34.