lp: Fix unbounded property when unboundedness has not been determined
Both Cplex and Gurobi have a status code that indicates that the problem
is either infeasible or unbounded. In this case, an additional
optimization is needed to determine if the problem is truly unbounded.
This implements the additional optimization when the unbounded property
is accessed so that is is only done when necessary.