boolean: Use Expression object as a unified representation
Previously the AND and OR operators were implemented with two separate
objects, duplicating a lot of code. In addition, the substitute method
could return one of couple of different types. This changes the Expression
to be a proper object that can be initialized from a string (parsed), a
variable, a boolean or an expression tree composed of And/Or objects. The
substitute method always returns an Expression object and if this object
was fully evaluated the bool can be obtained from the value property.

The terms now use a FrozenOrderedSet which means that the order of terms
will be preserved after parsing or substitution. I simplification of the
expression will still be performed though.