boolean: Provide the variables as a set
With this change the Expression object caches an immutable ordered set of
the variables in the expression. This means that creating the Expression
will check that the expression tree is correct and will build the set of
variables. The variables property (previously a method) then does not need
to iterate over the tree nodes and the return value also support the in
operator. This allows the caller to efficiently check if a variable is
present before trying to do substitution.