native: Only return model definition when present
Instead of parsing all reactions when the model definition is absent
the parse_model() method returns an empty iterator. The method
has_model_definition() is added to detect when this is the case.
This means that parsing the reactions can be avoided in the common case
where the model definition is not present.

The MetabolicMixin is changed to construct the MetabolicModel properly
with this change.