cplex: Reset problem type before solving
The previous solution to resetting the problem type for QP problems that
required global optimization did not work. Changing the problem type after
solving voids the solution, which meant that no solution was available in
this case. It also appeared that changing the objective of a QP problem
to a linear objective did not change the solver to LP.

To get around these issues, we simply change the problem type right before
calling solve based on the occurance of integer variables or quadratic
coefficients in the objective.