cplex: Work around segfault in Cplex library
Calling get_num_binary() on the Cplex instance appears to cause
a segfault in some rare cases. We work around this by avoiding
the call on the first solve. This may not work around the
segfault if the solve() method is called multiple times but at
least it works for problems that are only solved once.

The problem reset is only needed if a problem is modified after the
first solve, so it should be safe to skip the first part of
_reset_problem_type() if the problem has not been solved before.