kegg: Use abstract entry classes
This separates the raw KEGGEntry from the processed entries that
conform to the standard PSAMM properties. The raw entries are
returned from parse_kegg_entries(). These entries can then be
wrapped in kegg.CompoundEntry or kegg.ReactionEntry to expose
the properties following the PSAMM conventions. The mapping
is defined by two new classes CompoundMapper and ReactionMapper
that automatically map the raw properties and cache the result.