fastgapfill: Use compartment spec for extended model
When creating the extended model, use the compartment boundaries
specified in the model to generate transport reactions. For
consistency, this also changes the inclusion of database reactions
to only include reactions in the compartments that are specified
in the model. New warnings are output if no compartments are
specified in the model.