gapfind: Alternative GapFind algorithm which is more robust
Does not depend on v_max, instead sets a maximum flux when
solving this problem to 1. This reduces the distance between
v_max and epsilon making the problem more robust. This is
able to solve larger models that fail with the previous
implementation. This respects reaction directionality but
ignores any reaction flux limits. This also lowers the
maximum flux which scales the solution. This means that some
reactions may not be able to produce epsilon flux.