lp: Raise error on certain use of Expression/Relation
Raise an error when using chained comparisons with an
Expression object, e.g. 4 <= x <= 100 where x is an
Expression object. The "and" operator cannot be overridden
and currently it simply discards the first part of the
constraint because it evaluates it to True. Raising an
error when trying to evaluate an expression to bool
seems like the best way to catch this kind of mistake.