Use the name exchange internally for functions/variables
Changes variable and function names internally to use "exchange"
instead of "medium" or "media". This makes the use of this term
consistent with the file format and consistent with the fact that
this set of compounds specifies both allowed uptake ("the medium")
and also compounds that are allowed to leave the model system.