gapfilling: Fix performance issue when adding reactions
The functions for adding transport and exchange reactions would
supply the property model.reactions to the function that
generates a unique ID. This function expects a set in order to
do quick lookup but will also work with other collections that
support the `in` operator or even iterators. The performance issue
appeared for large models because model.reactions is not a set but
an iterator. This resulted in the gap-filling functions spending
large amounts of time traversing the same iterator over and over
when creating the extended model. This is fixed by first
creating a set from the iterator and using the set when
trying to generate unique IDs.