lp: Raise error on non-optimal solutions
This helps avoid a common bug where the result is not checked and
therefore non-optimal solution values are used as though an optimal
solution was found. Instead the solve() now raises a SolverError
when no optimal solution is found. This makes it impossible to
forget to check the solution in the most common case. In cases
when the exception is not desired the new solve_unchecked() can
be used which has the old behavior.