lp: Sort Product variables by text string
Previously these were sorted directly by the variable. Since the
variable is not necessarily an object that implements an ordering,
this will not work in Python 3. The sorting is only useful for pretty
printing the expression is converted to text in __str__. For now,
use text_type() as the sort key instead of the raw object.